Inlining all functions in Bin1D. Note that the only reason to not completely inline Bin2D is that there are some loops over bin edges which look a bit heavy to put into the header. But I think this is more than anything an artefact of the implementation in terms of all the edge points, including fourfold duplicated of each x/y coord, being very inefficient: Michal, can you rework or justify this structure, please? If the (x,y) edge points are *really* needed internally, let's calculate them on the fly from the minimal collection of 4 x/y low/high doubles unless there is a very compelling reason to do otherwise. I nearly swallowed my tongue when I worked out what that vector of pairs of pairs was for!

+++ src/Bin2D.cc
@@ -14,6 +14,7 @@ namespace YODA {
   Bin2D::Bin2D(double lowedgeX, double lowedgeY, double highedgeX, double highedgeY) {
     assert(lowedgeX <= highedgeX && lowedgeY <= highedgeY);
 
+    /// @todo Why store with so much redundancy?
     pair<pair<double,double>, pair<double,double> > edge1 =
       make_pair(make_pair(lowedgeX, lowedgeY), make_pair(lowedgeX, highedgeY));
     pair<pair<double,double>, pair<double,double> > edge2 =
@@ -23,21 +24,25 @@ namespace YODA {
     pair<pair<double,double>, pair<double,double> > edge4 =
       make_pair(make_pair(lowedgeX, lowedgeY), make_pair(highedgeX, lowedgeY));
 
-    _edges.push_back(edge1); _edges.push_back(edge2);
-    _edges.push_back(edge3); _edges.push_back(edge4);
+    _edges.push_back(edge1);
+    _edges.push_back(edge2);
+    _edges.push_back(edge3);
+    _edges.push_back(edge4);
   }
 
 
   Bin2D::Bin2D(std::vector<std::pair<std::pair<double,double>,
                std::pair<double,double> > > edges) {
     assert(edges.size() == 4);
-    _edges.push_back(edges[0]); _edges.push_back(edges[1]);
-    _edges.push_back(edges[2]); _edges.push_back(edges[3]);
+    _edges.push_back(edges[0]);
+    _edges.push_back(edges[1]);
+    _edges.push_back(edges[2]);
+    _edges.push_back(edges[3]);
   }
 
 
   void Bin2D::scaleXY(double scaleX, double scaleY) {
-    for (unsigned int i = 0; i < _edges.size(); ++i) {
+    for (size_t i = 0; i < _edges.size(); ++i) {
       _edges[i].first.first *= scaleX;
       _edges[i].second.first *= scaleX;
       _edges[i].first.second *= scaleY;
